move AnsweCallUseCase to usecase
